feat!: use pagination for sending large scans
The backend is introducing new required fields `page` and `totalPage`.

A new deployment is necessary on the cli side and the
backend to accommodate this breaking api change.

BREAKING CHANGE: introduce pagination
