+++ test/utils/mocks/scan-result-component.mock.ts
@@ -1,11 +1,12 @@
-import type { ScanResult, ScanResultComponent } from '../../../src/api/types/nes.types.ts';
+import type { InsightsEolScanComponent } from '../../../src/api/types/nes.types.ts';
+import type { ScanResult } from '../../../src/api/types/hd-cli.types.ts';
 
 export const createMockComponent = (
   purl: string,
   status: 'OK' | 'EOL' | 'LTS' = 'OK',
   eolAt: Date | null = null,
   daysEol: number | null = null,
-): ScanResultComponent => ({
+): InsightsEolScanComponent => ({
   purl,
   info: {
     eolAt,
@@ -16,7 +17,7 @@ export const createMockComponent = (
   },
 });
 
-export const createMockScan = (components: ScanResultComponent[]): ScanResult => ({
+export const createMockScan = (components: InsightsEolScanComponent[]): ScanResult => ({
   components: new Map(components.map((c) => [c.purl, c])),
   message: 'Test scan',
   success: true,
